Rename stores to persisted_store
