+++ src/lib/index.ts
@@ -1,4 +1,4 @@
 // place files you want to import through the `$lib` alias in this folder.
 export * from "./types"
-export * from "./stores"
+export * from "./persisted_store"
 export * from "./transitions"
